Remove GGL_SDK_EXPORT_API
This makes public functions always marked as default visibility instead
of depending on GGL_SDK_EXPORT_API. This doesn't create a size
difference in resulting binaries (though in the not-exported case, makes
function calls across complilation units go through the PLT). This does
mean that the symbols are always exported, but this can be addressed by
removing the default visibility attributes and -fvisbility=hidden,
allowing adding -fvisbility=hidden to the build when the user wants to
not export sdk symbols.